Remove duplicate code
[no ci]
