+++ test/integration/src/hictk_integration_suite/tests/balance.py
@@ -14,8 +14,10 @@ from hictk_integration_suite.runners.cooler import CoolerBalance
 from hictk_integration_suite.runners.hictk import HictkTestHarness
 from hictk_integration_suite.validators.file_formats import is_hic
 
+from .cli import HictkCli
 
-class _HictkBalanceCli(HictkTestHarness):
+
+class _HictkBalanceCli(HictkCli):
     def __init__(
         self,
         hictk_exec: pathlib.Path | str,
@@ -32,23 +34,6 @@ class _HictkBalanceCli(HictkTestHarness):
 
     def __repr__(self) -> str:
         return f"hictk-balance-{self._algorithm}-cli"
-
-    def _validate(self, expect_failure: bool):
-        if expect_failure:
-            if len(self.stderr()) == 0:
-                self._failures["missing help message"] = ""
-            if len(self.stdout()) != 0:
-                self._failures["unexpected output on stdout"] = self.stdout(500).strip()
-            if self.returncode == 0:
-                self._failures["unexpected return code"] = f"expected non-zero, found {self.returncode}"
-            return
-
-        if len(self.stdout()) == 0:
-            self._failures["missing help message"] = ""
-        if len(self.stderr()) != 0:
-            self._failures["unexpected output on stderr"] = self.stderr(500).strip()
-        elif not expect_failure and self.returncode != 0:
-            self._failures["unexpected return code"] = f"expected zero, found {self.returncode}"
 
 
 class _HictkBalance(HictkTestHarness):
