+++ test/integration/src/hictk_integration_suite/tests/cli.py
@@ -1,0 +1,35 @@
+# Copyright (C) 2024 Roberto Rossini <roberros@uio.no>
+#
+# SPDX-License-Identifier: MIT
+
+import json
+import logging
+import tomllib
+from timeit import default_timer as timer
+from typing import Any, Dict, List
+
+import yaml
+
+from hictk_integration_suite.runners.hictk import HictkTestHarness
+
+
+class HictkCli(HictkTestHarness):
+    def __repr__(self) -> str:
+        return "hictk-cli"
+
+    def _validate(self, expect_failure: bool):
+        if expect_failure:
+            if len(self.stderr()) == 0:
+                self._failures["missing help message"] = ""
+            if len(self.stdout()) != 0:
+                self._failures["unexpected output on stdout"] = self.stdout(500).strip()
+            if self.returncode == 0:
+                self._failures["unexpected return code"] = f"expected non-zero, found {self.returncode}"
+            return
+
+        if len(self.stdout()) == 0:
+            self._failures["missing help message"] = ""
+        if len(self.stderr()) != 0:
+            self._failures["unexpected output on stderr"] = self.stderr(500).strip()
+        elif not expect_failure and self.returncode != 0:
+            self._failures["unexpected return code"] = f"expected zero, found {self.returncode}"
