+++ test/integration/src/hictk_integration_suite/tests/dump.py
@@ -11,27 +11,12 @@ from hictk_integration_suite import validators
 from hictk_integration_suite.runners.common import normalize_df_dtypes
 from hictk_integration_suite.runners.hictk import HictkTestHarness
 
+from .cli import HictkCli
 
-class HictkDumpCli(HictkTestHarness):
+
+class HictkDumpCli(HictkCli):
     def __repr__(self) -> str:
-        return "hictk-metadata-dump"
-
-    def _validate(self, expect_failure: bool):
-        if expect_failure:
-            if len(self.stderr()) == 0:
-                self._failures["missing help message"] = ""
-            if len(self.stdout()) != 0:
-                self._failures["unexpected output on stdout"] = self.stdout(500).strip()
-            if expect_failure and self.returncode == 0:
-                self._failures["unexpected return code"] = f"expected non-zero, found {self.returncode}"
-            return
-
-        if len(self.stdout()) == 0:
-            self._failures["missing help message"] = ""
-        if len(self.stderr()) != 0:
-            self._failures["unexpected output on stderr"] = self.stderr(500).strip()
-        elif not expect_failure and self.returncode != 0:
-            self._failures["unexpected return code"] = f"expected zero, found {self.returncode}"
+        return "hictk-dump"
 
 
 class HictkDump(HictkTestHarness):
