+++ test/integration/src/hictk_integration_suite/tests/metadata.py
@@ -13,27 +13,12 @@ import yaml
 from hictk_integration_suite import validators
 from hictk_integration_suite.runners.hictk import HictkTestHarness
 
+from .cli import HictkCli
 
-class HictkMetadataCli(HictkTestHarness):
+
+class HictkMetadataCli(HictkCli):
     def __repr__(self) -> str:
         return "hictk-metadata-cli"
-
-    def _validate(self, expect_failure: bool):
-        if expect_failure:
-            if len(self.stderr()) == 0:
-                self._failures["missing help message"] = ""
-            if len(self.stdout()) != 0:
-                self._failures["unexpected output on stdout"] = self.stdout(500).strip()
-            if self.returncode == 0:
-                self._failures["unexpected return code"] = f"expected non-zero, found {self.returncode}"
-            return
-
-        if len(self.stdout()) == 0:
-            self._failures["missing help message"] = ""
-        if len(self.stderr()) != 0:
-            self._failures["unexpected output on stderr"] = self.stderr(500).strip()
-        elif not expect_failure and self.returncode != 0:
-            self._failures["unexpected return code"] = f"expected zero, found {self.returncode}"
 
 
 class HictkMetadata(HictkTestHarness):
