+++ test/integration/src/hictk_integration_suite/tests/zoomify.py
@@ -12,27 +12,12 @@ from hictk_integration_suite.runners.cooler import CoolerCoarsen
 from hictk_integration_suite.runners.hictk import HictkTestHarness
 from hictk_integration_suite.validators.file_formats import check_format, is_cooler
 
+from .cli import HictkCli
 
-class HictkZoomifyCli(HictkTestHarness):
+
+class HictkZoomifyCli(HictkCli):
     def __repr__(self) -> str:
         return "hictk-zoomify-cli"
-
-    def _validate(self, expect_failure: bool):
-        if expect_failure:
-            if len(self.stderr()) == 0:
-                self._failures["missing help message"] = ""
-            if len(self.stdout()) != 0:
-                self._failures["unexpected output on stdout"] = self.stdout(500).strip()
-            if self.returncode == 0:
-                self._failures["unexpected return code"] = f"expected non-zero, found {self.returncode}"
-            return
-
-        if len(self.stdout()) == 0:
-            self._failures["missing help message"] = ""
-        if len(self.stderr()) != 0:
-            self._failures["unexpected output on stderr"] = self.stderr(500).strip()
-        elif not expect_failure and self.returncode != 0:
-            self._failures["unexpected return code"] = f"expected zero, found {self.returncode}"
 
 
 class HictkZoomify(HictkTestHarness):
